ACE not initialising rule manager in constructor.
This was removed because ACE is lazy loaded in SE and there was a race between rule manager initialisation and actual rules reading in ACE.isAccessAllowed method. This resulted in blocking one channel complately (rebot required to unblock). We need to figure out order of components init if we wan't ACE to read rules at boot time.